Updated folders to contain README files.
Removed the place holders and replaced them with README.md (markdown) files with description of each folder. The descriptions are based on the information provided in the hongfire thread.

I also updated the root README and added most information provided in the thread, and changed it to markdown to make it slightly more prettier (or at least enable it).